+++ backend/src/main/java/sk/uniza/fri/telemedicine/repository/MeasurementPlanRepository.java
@@ -6,7 +6,7 @@ import sk.uniza.fri.telemedicine.entities.MeasurementPlan;
 
 import java.util.Optional;
 
-public interface MeasurementPlanRepository extends JpaRepository<MeasurementPlan, Integer> {
+public interface MeasurementPlanRepository extends JpaRepository<MeasurementPlan, Long> {
 
     @Query("SELECT m FROM MeasurementPlan m WHERE m.patient.personalNumber = :personalNumber AND m.validTo IS NULL")
     Optional<MeasurementPlan> findActivePlanByPersonalNumber(String personalNumber);
