+++ backend/src/main/java/sk/uniza/fri/telemedicine/repository/MeasurementTypePlanRepository.java
@@ -6,11 +6,11 @@ import sk.uniza.fri.telemedicine.entities.MeasurementTypePlan;
 
 import java.util.List;
 
-public interface MeasurementTypePlanRepository extends JpaRepository<MeasurementTypePlan, Integer> {
+public interface MeasurementTypePlanRepository extends JpaRepository<MeasurementTypePlan, Long> {
 
     @Query("SELECT mType FROM MeasurementTypePlan mType WHERE mType.measurementPlan.planId = :planId")
-    List<MeasurementTypePlan> findAllByPlanId(Integer planId);
+    List<MeasurementTypePlan> findAllByPlanId(Long planId);
 
     @Query("SELECT COUNT(mtp) > 0 FROM MeasurementTypePlan mtp WHERE mtp.measurementPlan.patient.personalNumber = :personalNumber AND mtp.measurementPlan.validTo IS NULL AND mtp.typeOfMeasurement.typeId = :typeId")
-    boolean existsByActivePlanAndTypeId(String personalNumber, Integer typeId);
+    boolean existsByActivePlanAndTypeId(String personalNumber, Long typeId);
 }
